refactoring shared logic for adding package json scriots for FF and FE
