+++ packages/fiori-elements-writer/src/index.ts
@@ -5,7 +5,7 @@ import type { App, Package } from '@sap-ux/ui5-application-writer';
 import { generate as generateUi5Project } from '@sap-ux/ui5-application-writer';
 import { generate as addOdataService, OdataVersion, ServiceType } from '@sap-ux/odata-service-writer';
 import { generateOPAFiles } from '@sap-ux/ui5-test-writer';
-import { getPackageJsonTasks } from './packageConfig';
+import { generatePackageScripts } from '@sap-ux/fiori-generator-shared';
 import cloneDeep from 'lodash/cloneDeep';
 import type { FioriElementsApp } from './types';
 import { TemplateType } from './types';
@@ -177,7 +177,7 @@ async function generate<T extends {}>(basePath: string, data: FioriElementsApp<T
     if (isEdmxProjectType) {
         // Add scripts to package.json only for non-CAP projects
         packageJson.scripts = Object.assign(packageJson.scripts ?? {}, {
-            ...getPackageJsonTasks({
+            ...generatePackageScripts({
                 localOnly: !feApp.service?.url,
                 addMock: !!feApp.service?.metadata,
                 addTest,
