+++ packages/fiori-freestyle-writer/src/translations/fiori-freestyle-writer.i18n.json
@@ -1,5 +1,4 @@
 {
     "info": {
-        "mockOnlyWarning": "This application was generated with a local metadata file and does not reference a live server. Please add the required server configuration or start this application with mock data using the target: npm run start-mock"
     }
 }
