+++ packages/fiori-generator-shared/src/index.ts
@@ -1,5 +1,6 @@
 export * from './cap';
 export * from './environment';
 export { getBootstrapResourceUrls } from './helpers';
+export { generatePackageScripts } from './generatePackageScripts';
 export { generateReadMe } from './read-me';
 export * from './system-utils';
