+++ packages/fiori-generator-shared/src/translations/fiori-generator-shared.i18n.json
@@ -1,0 +1,5 @@
+{
+    "info": {
+        "mockOnlyWarning": "This application was generated with a local metadata file and does not reference a live server. Please add the required server configuration or start this application with mock data using the target: npm run start-mock"
+    }
+}
