+++ packages/fiori-generator-shared/src/types.ts
@@ -54,3 +54,14 @@ export interface ReadMe {
     /** Additional custom entries for the application. */
     additionalEntries?: AdditionalEntries[];
 }
+
+/**
+Represents a set of scripts defined in a package.json file.
+ */
+export interface PackageJsonScripts {
+    start: string;
+    'start-local': string;
+    'start-noflp'?: string;
+    'start-mock'?: string;
+    'int-test'?: string;
+}
